Polyfill moved to module (core-js only supports ECMAscript features)
