+++ CancerGov/_src/Scripts/NCI/UX/Common/polyfills/custom_event.js
@@ -1,0 +1,15 @@
+(function () {
+    if ( typeof window.CustomEvent === "function" ) return false;
+    
+    function CustomEvent ( event, params ) {
+        console.log('Polyfilling CustomEvent')
+        params = params || { bubbles: false, cancelable: false, detail: undefined };
+        var evt = document.createEvent( 'CustomEvent' );
+        evt.initCustomEvent( event, params.bubbles, params.cancelable, params.detail );
+        return evt;
+     }
+  
+    CustomEvent.prototype = window.Event.prototype;
+  
+    window.CustomEvent = CustomEvent;
+})();
